typescriptifying InviteTileViewModel
